requirements.txt
accelerate==0.31.0
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.2
altair==4.2.2
annotated-types==0.7.0
anyio==3.7.1
appdirs==1.4.4
asgiref==3.8.1
attrs==23.2.0
audio-recorder-streamlit==0.0.9
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.34.136
botocore==1.34.136
Brotli==1.1.0
build==1.2.1
cachetools==5.3.3
certifi==2024.6.2
chardet==3.0.4
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
cohere==5.5.8
coloredlogs==15.0.1
contourpy==1.2.1
crewai==0.35.0
cycler==0.12.1
dataclasses-json==0.6.7
deep-translator==1.11.4
deepdiff==7.0.1
Deprecated==1.2.14
distro==1.9.0
dnspython==2.6.1
docstring_parser==0.16
duckduckgo_search==5.3.1b1
email_validator==2.2.0
embedchain==0.1.109
emoji==2.12.1
entrypoints==0.4
Faker==26.0.0
fastapi==0.111.0
fastapi-cli==0.0.4
fastavro==1.9.4
favicon==0.7.0
ffmpeg-python==0.2.0
filetype==1.2.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.0
frozenlist==1.4.1
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.1
google-auth==2.30.0
google-cloud-aiplatform==1.57.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
gptcache==0.1.43
greenlet==3.0.3
groq==0.9.0
grpc-google-iam-v1==0.13.1
grpcio==1.64.1
grpcio-status==1.62.2
gTTS==2.5.1
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.7.1
htbuilder==0.6.2
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.23.4
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
importlib-metadata==6.11.0
importlib_resources==6.4.0
instructor==1.3.3
intel-openmp==2021.4.0
itsdangerous==2.2.0
jiter==0.4.2
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
kubernetes==30.1.0
langchain==0.1.20
langchain-chroma==0.1.2
langchain-cohere==0.1.5
langchain-community==0.0.38
langchain-core==0.2.10
langchain-groq==0.1.6
langchain-openai==0.1.7
langchain-text-splitters==0.0.2
langdetect==1.0.9
langgraph==0.1.4
langsmith==0.1.82
llvmlite==0.43.0
lxml==5.2.2
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
markdownlit==0.0.7
marshmallow==3.21.3
matplotlib==3.9.1
mdurl==0.1.2
mkl==2021.4.0
mkl-service==2.4.0
mmh3==4.1.0
monotonic==1.6
more-itertools==10.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nltk==3.8.1
numba==0.60.0
oauthlib==3.2.2
onnxruntime==1.18.1
openai==1.35.7
openai-whisper==20231117
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
ordered-set==4.1.0
orjson==3.10.5
overrides==7.7.0
packaging==23.2
pandas==2.2.2
parameterized==0.9.0
pillow==10.2.0
posthog==3.5.0
prometheus_client==0.20.0
proto-plus==1.24.0
protobuf==4.25.3
pulsar-client==3.5.0
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
pydantic==2.7.4
pydantic_core==2.18.4
pydeck==0.9.1
pydub==0.25.1
pymdown-extensions==10.8.1
pyparsing==3.1.2
pypdf==4.2.0
PyPika==0.48.9
pyproject_hooks==1.1.0
pyreadline3==3.4.1
pyreqwest_impersonate==0.4.9
pysbd==0.3.4
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-multipart==0.0.9
pytz==2024.1
pywin32==305.1
PyYAML==6.0.1
rapidfuzz==3.9.3
referencing==0.35.1
regex==2023.12.25
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
s3transfer==0.10.2
safetensors==0.4.3
schema==0.7.7
scikit-learn==1.5.0
scipy==1.14.0
sentence-transformers==3.0.1
setuptools==69.5.1
shapely==2.0.4
shellingham==1.5.4
smmap==5.0.1
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.5
SpeechRecognition==3.10.4
SQLAlchemy==2.0.31
st-annotated-text==4.0.1
st-multimodal-chatinput==0.1.6
st-theme==1.2.3
starlette==0.37.2
streamlit==1.36.0
streamlit-audiorec==0.1.3
streamlit-audiorecorder==0.0.5
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-chat==0.0.2.2
streamlit-embedcode==0.1.2
streamlit-extras==0.4.3
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.2.4
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
streamlit_mic_recorder==0.0.8
streamlit_TTS==0.0.7
sympy==1.12.1
tabulate==0.9.0
tbb==2021.13.0
tenacity==8.4.2
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.3.1+cu121
torchaudio==2.3.1+cu121
torchvision==0.18.1+cu121
tqdm==4.66.4
transformers==4.42.3
typer==0.12.3
types-requests==2.32.0.20240622
typing-inspect==0.9.0
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
unstructured==0.14.9
unstructured-client==0.23.8
urllib3==2.2.2
uvicorn==0.30.1
validators==0.30.0
watchdog==4.0.1
watchfiles==0.22.0
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
yarl==1.9.4